Update upload.php - Preventative security fixes [UNTESTED]
This is a very quick and untested push, will test and update as required.

I saw a ShareX uploader bruteforce script some time ago which targetted uploaders like this with common credentials and missing filetype validation. Pushing this update to hopefully begin to solve this problem.

The thing with filetype validation and this particular script is that I recall having validation before, correct me if I am wrong but it must have been removed sometime between December 2019 and my last commit (December 2020?) - If not I've been on too much.

I don't know how big of an issue this actually is, obviously isn't news breaking and I haven't personally seen any instances of it happening here so better to be safe than sorry.